requirements.txt
Flask
gradio
ibm-watson-machine-learning
openai-whisper
python-dotenv
gTTS
